chore(demos): remove old demos.prod file and rename new one to it
