+++ scripts/gulp/gulpfile.ts
@@ -3,8 +3,7 @@ import './tasks/clean';
 import './tasks/default';
 import './tasks/demos';
 import './tasks/demos.dev';
-// import './tasks/demos.prod';
-import './tasks/demos.scripts';
+import './tasks/demos.prod';
 import './tasks/docs';
 import './tasks/e2e';
 import './tasks/e2e.dev';
