+++ scripts/gulp/tasks/demos.prod.ts
@@ -1,38 +1,44 @@
-import { accessSync, F_OK, readFileSync, writeFileSync } from 'fs';
+import { spawnSync } from 'child_process';
+import { accessSync, F_OK, readdirSync, readFileSync, stat, statSync } from 'fs';
 import { dirname, join } from 'path';
 
-import * as glob from 'glob';
 import { dest, src, start, task } from 'gulp';
 import * as gulpif from 'gulp-if';
 import * as watch from 'gulp-watch';
-import * as rollup from 'rollup';
-import * as nodeResolve from 'rollup-plugin-node-resolve';
-import * as commonjs from 'rollup-plugin-commonjs';
 import * as runSequence from 'run-sequence';
 import { obj } from 'through2';
 import * as VinylFile from 'vinyl';
-import { argv } from 'yargs';
 
-import { DEMOS_SRC_ROOT, DIST_DEMOS_COMPONENTS_ROOT, DIST_DEMOS_ROOT, DIST_NAME, DEMOS_NAME, ES5, ES_2015, LOCAL_SERVER_PORT, PROJECT_ROOT, SCRIPTS_ROOT, SRC_ROOT } from '../constants';
-import { createTempTsConfig, deleteFiles, getFolderInfo, runNgc } from '../util';
+import { DEMOS_SRC_ROOT, DIST_DEMOS_ROOT, DIST_NAME, DEMOS_NAME, ES5, ES_2015, LOCAL_SERVER_PORT, SCRIPTS_ROOT } from '../constants';
+import { createTempTsConfig, getFolderInfo } from '../util';
 
 task('demos.prod', demosBuild);
 
 function demosBuild(done: (err: any) => void) {
   runSequence(
+    'demos.copyIonic',
     'demos.clean',
     'demos.polyfill',
     'demos.copySource',
-    'demos.compileTests',
     'demos.copyExternalDependencies',
     'demos.sass',
     'demos.fonts',
-    'demos.bundleProd',
+    'demos.compileTests',
     done);
 }
 
+task('demos.copyIonic', (done: (err: any) => void) => {
+  runSequence(
+    'compile.release',
+    'release.compileSass',
+    'release.fonts',
+    'release.sass',
+    'release.createUmdBundle',
+    done);
+});
+
 task('demos.copySource', (done: Function) => {
-  const stream = src([`${SRC_ROOT}/**/*`, `!${SRC_ROOT}/components/*/test{,/**/*}`, `${DEMOS_SRC_ROOT}/**/*`])
+  const stream = src([`${DEMOS_SRC_ROOT}/**/*`])
     .pipe(gulpif(/app.module.ts$/, createIndexHTML()))
     .pipe(dest(DIST_DEMOS_ROOT));
 
@@ -40,7 +46,7 @@ task('demos.copySource', (done: Function) => {
 
   function createIndexHTML() {
     const indexTemplate = readFileSync(`${SCRIPTS_ROOT}/${DEMOS_NAME}/demos.template.prod.html`);
-    const indexTs = readFileSync(`${SCRIPTS_ROOT}/${DEMOS_NAME}/entry.ts`);
+    const indexTs = readFileSync(`${SCRIPTS_ROOT}/${DEMOS_NAME}/main.ts`);
 
     return obj(function (file, enc, next) {
       this.push(new VinylFile({
@@ -51,7 +57,7 @@ task('demos.copySource', (done: Function) => {
       this.push(new VinylFile({
         base: file.base,
         contents: new Buffer(indexTs),
-        path: join(dirname(file.path), 'entry.ts'),
+        path: join(dirname(file.path), 'main.ts'),
       }));
       next(null, file);
     });
@@ -60,104 +66,82 @@ task('demos.copySource', (done: Function) => {
 
 task('demos.compileTests', (done: Function) => {
   let folderInfo = getFolderInfo();
-  buildDemoTests(folderInfo, done);
-});
 
-function buildDemoTests(folderInfo: any, done: Function) {
-  let includeGlob = ['./*/app.module.ts', './*/entry.ts'];
   if (folderInfo.componentName && folderInfo.componentTest) {
-    includeGlob = [
-      `./${folderInfo.componentName}/app.module.ts`,
-      `./${folderInfo.componentName}/entry.ts`,
-    ];
-  }
-
-  createTempTsConfig(includeGlob, ES5, ES_2015, `${PROJECT_ROOT}/tsconfig.json`, `${DIST_DEMOS_ROOT}/tsconfig.json`);
-  runNgc(`${DIST_DEMOS_ROOT}/tsconfig.json`, (err) => {
-    if (err) {
-      done(err);
-      return;
-    }
-    // clean up any .ts files that remain
-    deleteFiles([`${DIST_DEMOS_ROOT}/**/*.ts`, `!${DIST_DEMOS_ROOT}/**/*.ngfactory.ts`, `!${DIST_DEMOS_ROOT}/**/*.d.ts`], done);
-  });
-}
-
-task('demos.bundleProd', (done) => {
-  let includeGlob = `${DIST_DEMOS_ROOT}/*/entry.js`;
-  let folderInfo = getFolderInfo();
-  if (folderInfo.componentName && folderInfo.componentTest) {
-    includeGlob = `${DIST_DEMOS_ROOT}/${folderInfo.componentName}/entry.js`;
-  }
-  glob(includeGlob, {}, function (er, files) {
-    var directories = files.map(function (file) {
-      return dirname(file);
-    });
-
-    let indexFileContents = directories.map(function (dir) {
-      let testName = dir.replace(`${DIST_DEMOS_ROOT}/`, '');
-      let fileName = dir.replace(`${PROJECT_ROOT}`, '');
-      return `<p><a href="${fileName}">${testName}</a></p>`;
-    }, []);
-
-    writeFileSync(`${DIST_DEMOS_ROOT}/index.html`,
-      '<!DOCTYPE html><html lang="en"><head></head><body style="width: 500px; margin: 100px auto">\n' +
-      indexFileContents.join('\n') +
-      '</center></body></html>'
-    );
-
-    createBundles(files).then(() => {
-      done();
-    }).catch(err => {
-      done(err);
-    });
-  });
-});
-
-function createBundles(files: string[]) {
-  let start;
-  if (!files) {
-    return Promise.reject(new Error('list of files is null'));
-  } else if (files.length === 0) {
-    return Promise.resolve();
+    buildTest(folderInfo.componentName);
   } else {
-    const outputFileName = join(dirname(files[0]), 'app.bundle.js');
-    start = Date.now();
-    return bundle(files[0], outputFileName).then(() => {
-      const end = Date.now();
-      const seconds = (end - start) / 1000;
-      console.log(`Took ${seconds} seconds to process ${files[0]}`);
-      const remainingFiles = files.concat();
-      remainingFiles.shift();
-      return createBundles(remainingFiles);
-    }).catch(err => {
-      return Promise.reject(err);
-    });
+    buildAllTests(done);
   }
+});
+
+function buildTest(folderName: string) {
+  let includeGlob = [`./dist/demos/${folderName}/*.ts`];
+  let pathToWriteFile = `${DIST_DEMOS_ROOT}/${folderName}/tsconfig.json`;
+
+  createTempTsConfig(includeGlob, ES5, ES_2015, `${DEMOS_SRC_ROOT}/tsconfig.json`, pathToWriteFile);
+  return runAppScripts(folderName);
 }
 
-function bundle(inputFile: string, outputFile: string): Promise<any> {
-  console.log(`Starting rollup on ${inputFile} ... writing to ${outputFile}`);
-  return rollup.rollup({
-    entry: inputFile,
-    plugins: [
-      commonjs(),
-      nodeResolve({
-        module: true,
-        jsnext: true,
-        main: true,
-        extensions: ['.js']
-      })
-    ]
-  }).then(bundle => {
-    return bundle.write({
-      format: 'iife',
-      dest: outputFile,
+function buildAllTests(done: Function) {
+  let folders = getFolders('./dist/demos/');
+  let promises: Promise<any>[] = [];
+
+  folders.forEach(folder => {
+    stat(`./dist/demos/${folder}/app.module.ts`, function(err, stat) {
+      if (err == null) {
+        const promise = buildTest(folder);
+        promises.push(promise);
+      }
     });
+  });
+
+  Promise.all(promises).then(() => {
+    done();
+  }).catch(err => {
+    done(err);
   });
 }
 
-task('demos.watchProd', ['demos.copyExternalDependencies', 'demos.sass', 'demos.fonts'], (done: Function) => {
+function runAppScripts(folderName: string) {
+  console.log('Running app scripts with', folderName);
+
+  let sassConfigPath = 'scripts/demos/sass.config.js';
+
+  let appEntryPoint = `dist/demos/${folderName}/main.ts`;
+  let distDir = `dist/demos/${folderName}/`;
+
+  let tsConfig = distDir + 'tsconfig.json';
+
+  try {
+    const scriptsCmd = spawnSync('ionic-app-scripts',
+      ['build',
+      '--prod',
+      '--sass', sassConfigPath,
+      '--appEntryPoint', appEntryPoint,
+      '--srcDir', distDir,
+      '--wwwDir', distDir,
+      '--tsconfig', tsConfig
+      ]);
+
+    if (scriptsCmd.status !== 0) {
+      return Promise.reject(scriptsCmd.stderr.toString());
+    }
+
+    console.log(scriptsCmd.output.toString());
+    return Promise.resolve();
+  } catch (ex) {
+    return Promise.reject(ex);
+  }
+}
+
+function getFolders(dir) {
+  return readdirSync(dir)
+    .filter(function(file) {
+      return statSync(join(dir, file)).isDirectory();
+    });
+}
+
+task('demos.watchProd', (done: Function) => {
   const folderInfo = getFolderInfo();
   let demoTestPath = DEMOS_SRC_ROOT;
 
@@ -172,13 +156,13 @@ task('demos.watchProd', ['demos.copyExternalDependencies', 'demos.sass', 'demos.
     return;
   }
 
-  if (demosComponentsExists()) {
+  if (demosComponentsExists(folderInfo)) {
     // already generated the demos directory
     demosWatch(folderInfo.componentName, folderInfo.componentTest);
 
   } else {
     // generate the demos directory
-    console.log('Generated demos builds first...');
+    console.log('Generate demo builds first...');
     demosBuild(() => {
       demosWatch(folderInfo.componentName, folderInfo.componentTest);
     });
@@ -217,9 +201,15 @@ function demosWatch(componentName: string, componentTest: string) {
   start('demos.serve');
 }
 
-function demosComponentsExists(): boolean {
+function demosComponentsExists(folderInfo: any): boolean {
+  let componentPath = DIST_DEMOS_ROOT;
+
+  if (folderInfo.componentName && folderInfo.componentTest) {
+    componentPath += `/${folderInfo.componentName}/build`;
+  }
+
   try {
-    accessSync(DIST_DEMOS_COMPONENTS_ROOT, F_OK);
+    accessSync(componentPath, F_OK);
   } catch (e) {
     return false;
   }
